+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 Significant changes in this project will be documented here.
 
+### [1.0.6] - 2018-12-16
+
+There are some small breaking changes in this version.
+
+### Changed
+
+- `DateTimeSpecification` is now `DateSpecification`
+- The `DateTime` extension method `.Spec` is now `.ComplyWith`
+- The `Exception` extension methods related to SQL Server is now strongly-typed associated to `SqlException` class.
+
+### Added
+
+- New `String` extension methods to do convertions to `float, sbyte, uint, ulong` and `ushort`
+- `DataRow` extension methods, `FieldAs<T>`.
+
 ### [1.0.5] - 2018-10-20
 
 ### Changed